increase spawnId width from uint32 to uint64

+++ sql/updates/world/master/2024_xx_xx_xx_world.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS `creature_static_flags_override`;
 CREATE TABLE `creature_static_flags_override` (  
-  `SpawnId` INT UNSIGNED NOT NULL,
+  `SpawnId` BIGINT UNSIGNED NOT NULL,
   `DifficultyId` TINYINT UNSIGNED NOT NULL DEFAULT 0,
   `StaticFlags1` INT UNSIGNED,
   `StaticFlags2` INT UNSIGNED,
